apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
httpx==0.27.2
redis==5.1.0
celery==5.4.0
arq==0.26.1
boto3==1.35.36
aioboto3==13.2.0
python-multipart==0.0.12
Pillow==10.4.0
weasyprint==62.3
email-validator==2.2.0
pgvector==0.3.5
python-dotenv==1.0.1
# AI
langchain==0.3.1
langgraph==0.2.28
langchain-anthropic==0.2.1
langchain-openai==0.2.0
anthropic==0.39.0
openai==1.51.0
